opciones de administrador v2

+++ view/controlCalidad/indexTemplate.html.php
@@ -5,6 +5,8 @@ use mvc\routing\routingClass as routing ?>
 use mvc\i18n\i18nClass as i18n ?>
 <?php
 use mvc\view\viewClass as view ?>
+<?php
+use mvc\session\sessionClass as session ?>
 <?php $id = controlCalidadTableClass::ID ?>
 <?php $fecha = controlCalidadTableClass::FECHA ?>
 <?php $turno = controlCalidadTableClass::TURNO ?>
@@ -194,8 +196,10 @@ use mvc\view\viewClass as view ?>
 
     <form id="frmDeleteAll" action="<?php echo routing::getInstance()->getUrlWeb('controlCalidad', 'deleteSelect') ?>" method="POST">
         <div style="margin-bottom: 10px; margin-top: 30px">
-            <a href="<?php echo routing::getInstance()->getUrlWeb('controlCalidad', 'insert') ?>" class="btn btn-success btn-xs"><?php echo i18n::__('new') ?></a>
-            <a href="javascript:eliminarMasivo()" class="btn btn-danger btn-xs" id="btnDeleteMass" data-toggle="modal" data-target="#myModalDeleteMass"><?php echo i18n::__('deleteSelect') ?></a>
+            <?php if (session::getInstance()->hasCredential('admin')): ?> 
+                <a href="<?php echo routing::getInstance()->getUrlWeb('controlCalidad', 'insert') ?>" class="btn btn-success btn-xs"><?php echo i18n::__('new') ?></a>
+                <a href="javascript:eliminarMasivo()" class="btn btn-danger btn-xs" id="btnDeleteMass" data-toggle="modal" data-target="#myModalDeleteMass"><?php echo i18n::__('deleteSelect') ?></a>
+            <?php endif; ?>
             <button type="button" data-toggle="modal" data-target="#myModalFilters" class="btn btn-primary  btn-xs"><?php echo i18n::__('filters') ?></button>
             <a href="<?php echo routing::getInstance()->getUrlWeb('controlCalidad', 'deleteFilters') ?>" class="btn btn-default btn-xs"><?php echo i18n::__('deleteFilters') ?></a>
             <a class="btn btn-warning btn-xs col-lg-offset-7" data-toggle="modal" data-target="#myModalFILTROSREPORTE" ><?php echo i18n::__('printReport') ?></a>
@@ -213,34 +217,36 @@ use mvc\view\viewClass as view ?>
             </thead>
             <tbody>
                 <?php foreach ($objControlCalidad as $control): ?>
-                  <tr>
-                      <td><input type="checkbox" name="chk[]" value="<?php echo $control->$id ?>"></td>
-                      <td><?php echo $control->$fecha ?></td>
-                      <td><?php echo controlCalidadTableClass::getNameEmpleado($control->$empleado_id) ?></td>
-                      <td><?php echo controlCalidadTableClass::getNameProveedor($control->$proveedor_id) ?></td>
-                      <td>
-                          <a href="<?php echo routing::getInstance()->getUrlWeb('controlCalidad', 'view', array(controlCalidadTableClass::ID => $control->$id)) ?>" class="btn btn-info btn-xs"><?php echo i18n::__('view') ?></a></a>
-                          <a href="<?php echo routing::getInstance()->getUrlWeb('controlCalidad', 'edit', array(controlCalidadTableClass::ID => $control->$id)) ?>" class="btn btn-primary btn-xs"><?php echo i18n::__('edit') ?></a></a>
-                          <a href="#" data-toggle="modal" data-target="#myModalDelete<?php echo $control->$id ?>" class="btn btn-danger btn-xs"><?php echo i18n::__('delete') ?></a>
-                      </td>
-                  </tr>
-              <div class="modal fade" id="myModalDelete<?php echo $control->$id ?>" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
-                  <div class="modal-dialog">
-                      <div class="modal-content">
-                          <div class="modal-header">
-                              <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-                              <h4 class="modal-title" id="myModalLabel"><?php echo i18n::__('confirmDelete') ?></h4>
-                          </div>
-                          <div class="modal-body">
-                              <?php echo i18n::__('questionDelete') ?> <?php echo $control->$id ?>?
-                          </div>
-                          <div class="modal-footer">
-                              <button type="button" class="btn btn-default" data-dismiss="modal"><?php echo i18n::__('cancel') ?></button>
-                              <button type="button" class="btn btn-primary" onclick="eliminar(<?php echo $control->$id ?>, '<?php echo controlCalidadTableClass::getNameField(controlCalidadTableClass::ID, true) ?>', '<?php echo routing::getInstance()->getUrlWeb('controlCalidad', 'delete') ?>')"><?php echo i18n::__('confirmDelete') ?></button>
-                          </div>
-                      </div>
-                  </div>
-              </div>
+                    <tr>
+                        <td><input type="checkbox" name="chk[]" value="<?php echo $control->$id ?>"></td>
+                        <td><?php echo $control->$fecha ?></td>
+                        <td><?php echo controlCalidadTableClass::getNameEmpleado($control->$empleado_id) ?></td>
+                        <td><?php echo controlCalidadTableClass::getNameProveedor($control->$proveedor_id) ?></td>
+                        <td>
+                            <a href="<?php echo routing::getInstance()->getUrlWeb('controlCalidad', 'view', array(controlCalidadTableClass::ID => $control->$id)) ?>" class="btn btn-info btn-xs"><?php echo i18n::__('view') ?></a></a>
+                            <?php if (session::getInstance()->hasCredential('admin')): ?>
+                                <a href="<?php echo routing::getInstance()->getUrlWeb('controlCalidad', 'edit', array(controlCalidadTableClass::ID => $control->$id)) ?>" class="btn btn-primary btn-xs"><?php echo i18n::__('edit') ?></a></a>
+                                <a href="#" data-toggle="modal" data-target="#myModalDelete<?php echo $control->$id ?>" class="btn btn-danger btn-xs"><?php echo i18n::__('delete') ?></a>
+                            <?php endif; ?>
+                        </td>
+                    </tr>
+                <div class="modal fade" id="myModalDelete<?php echo $control->$id ?>" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
+                    <div class="modal-dialog">
+                        <div class="modal-content">
+                            <div class="modal-header">
+                                <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+                                <h4 class="modal-title" id="myModalLabel"><?php echo i18n::__('confirmDelete') ?></h4>
+                            </div>
+                            <div class="modal-body">
+                                <?php echo i18n::__('questionDelete') ?> <?php echo $control->$id ?>?
+                            </div>
+                            <div class="modal-footer">
+                                <button type="button" class="btn btn-default" data-dismiss="modal"><?php echo i18n::__('cancel') ?></button>
+                                <button type="button" class="btn btn-primary" onclick="eliminar(<?php echo $control->$id ?>, '<?php echo controlCalidadTableClass::getNameField(controlCalidadTableClass::ID, true) ?>', '<?php echo routing::getInstance()->getUrlWeb('controlCalidad', 'delete') ?>')"><?php echo i18n::__('confirmDelete') ?></button>
+                            </div>
+                        </div>
+                    </div>
+                </div>
             <?php endforeach ?>
             </tbody>
         </table>
@@ -248,7 +254,7 @@ use mvc\view\viewClass as view ?>
     <div class="text-right">
         <?php echo i18n::__('page') ?>  <select id="slqPaginador" onchange="paginador(this, '<?php echo routing::getInstance()->getUrlWeb('empleado', 'index') ?>')">
         <?php for ($x = 1; $x <= $cntPages; $x++): ?>
-              <option <?php echo(isset($page) and $page == $x) ? 'selected' : '' ?> value="<?php echo $x ?>"><?php echo $x ?></option>
+                <option <?php echo(isset($page) and $page == $x) ? 'selected' : '' ?> value="<?php echo $x ?>"><?php echo $x ?></option>
             <?php endfor ?>
         </select> <?php echo i18n::__('of') ?> <?php echo $cntPages ?>
     </div>
